Fix with_items statements for ansible 2.2
In ansible 2.2 (and maybe lower versions), it is necessary
that for lists from variables the syntax
with_items: '{{ variable }}'
is used. Otherwise operations like item.property try
to work on the complete list and not a single elements
which causes the error message that property "property" does
not exists on the element "item".